Allow to use in functional testing
Using stacked kernels within functional testing I got errors because the kernel was not shut down at the end of each request. 
This fix solved my issues, I'm not sure it is the only way (or the best way) to go though. I can't see any contraindication with this, 
but I have very little experience with stackPHP